refactor(tests): split test_outbox.py into 5 focused modules
Split the 1160-line test/adapters/driving/fastapi/test_outbox.py (47 tests)
into 5 files grouped by functional area:

- test_outbox_handler.py        (7 tests, 191 lines): FIFO loop, retry/abort,
                                  actor resolution
- test_outbox_handle_item_delivery.py (10 tests, 324 lines): delivery, logging,
                                  skip conditions, DR-01 regression,
                                  Announce(CaseLedgerEntry) inline fields
- test_outbox_handle_item_validation.py (17 tests, 297 lines): expansion bridge
                                  (P347-BRIDGE), OX-08 to: enforcement,
                                  OX-09 integrity errors
- test_outbox_helpers.py        (17 tests, 241 lines): _extract_recipients,
                                  _format_object, _dehydrate_references,
                                  _is_stub_object_dict, _coerce_reference_value
- test_outbox_object_pipeline.py (6 tests, 283 lines): _recover_typed_inline_object,
                                  _hydrate_inline_object_if_persistable,
                                  _prepare_activity_object_for_delivery,
                                  handle_outbox_item hydration integration

All 57 tests pass; no logic changes. Shared helpers are local to the files
that use them; _make_vultron_activity is duplicated where needed.

Closes #497
